requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pytesseract>=0.3.10
Pillow>=9.0.0
fpdf2>=2.7.0
python-docx>=0.8.11
PyMuPDF>=1.23.0

# AI packages for study guide generation (install as needed)
openai>=1.3.0
google-generativeai>=0.3.0
anthropic>=0.7.0
ollama>=0.1.7
pytesseract==0.3.10
pillow==10.0.1
fpdf2==2.7.6
python-docx==1.1.0